requirements.txt
anthropic>=0.40.0
python-telegram-bot>=21.0
chromadb>=0.5.0
pymupdf>=1.24.0
sentence-transformers>=3.0.0
python-dotenv>=1.0.0
aiosqlite>=0.20.0
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.0.0
supabase>=2.4.0
PyJWT>=2.8.0
